refactor: move searchBar and sortDropdown components under item directory
